Valor das páginas foi alterado de Long para Int.

+++ app/src/main/java/com/example/architecturecomponentapp/data/database/remote/FilmApiService.kt
@@ -34,12 +34,12 @@ interface FilmApiService {
 
     // chamar lista de filmes mais populares da API
     @GET("movie/popular${Api.UNIQUE_KEY}${Api.LANGUAGE}")
-    fun callPopularMovieListApi(@Query("page") page: Long): Deferred<FilmsJson>
+    fun callPopularMovieListApi(@Query("page") page: Int): Deferred<FilmsJson>
 
     // buscar lista de filmes na API pelo token search
     @GET("search/movie${Api.UNIQUE_KEY}${Api.LANGUAGE}")
     fun callSearchMovieList(
-        @Query("page") page: Long,
+        @Query("page") page: Int,
         @Query("query") search: String
     ): Deferred<FilmsJson>
